+++ src/pages/subPages/faq/faq.jsx
@@ -3,7 +3,7 @@ import { Fcontents, fReducer } from "./faqData"
 import { useMemo, useReducer } from "react"
 
 import SupTop from '../../../components/common/supTop'
-import OneDep from '../../../components/onedep'
+import { FaqDep } from '../fesInfo/infoDep'
 import FaqPagination from './faqpaginations'
 import FaqSearch from './faqSearch'
 
@@ -11,8 +11,6 @@ import FaqSearch from './faqSearch'
 const Faq = () => {
     const [state, dispatch] = useReducer(fReducer, Fcontents)
     const {fconts, fsearch} = state;
-    const OsubTitsF = ['자주하는 질문','묻고 답하기']
-    const TlinksF = ['/Faq']
 
     const Fsearch = (text) =>{
         dispatch({
@@ -27,10 +25,12 @@ const Faq = () => {
     
     const allCount = useMemo(()=>allCounts(fconts),[fconts])
 
+    const one = 1;
+
     return(
       <>
         <SupTop supTopImg={'subSupportTop'} supTopTit={'고객지원'}/>
-        <OneDep OsubTits={OsubTitsF} Olinks={TlinksF}/>
+        <FaqDep one={one}/>
         <section className="faqSec">
         <h6>total <span>{allCount}</span></h6>
         <span className='faqTopLine'></span>
