+++ src/pages/subPages/fesInfo/subFesInfoPages.jsx
@@ -1,38 +1,18 @@
 import { SubFesInfoK, SubFesInfoC, ProgramK, ProgramC, TicketK, TicketC } from "./subFesInfoCom"
 import { useState } from "react";
 import { Link } from 'react-router-dom';
-import '../../../components/onedep.css'
-import '../../../components/twodep.css'
 import SupTop from "../../../components/common/supTop"
 import LineTit from "../../../components/lineTit"
-import Scr
+import {InfoDep, InfoDep2, InfoDep3}  from "./infoDep";
 
 function Kinfo(){
 
-  const [oneDep, setOneDep] = useState(1);
-
-  const [twoDep, setTwoDep] = useState(1);
-    
-  function clickOne(index){
-      setOneDep(index);
-  }
-
-  function clickTwo(index){
-    setTwoDep(index);
-}
 
     return(
         <>
         <SupTop supTopImg={'subFesInfoImage'} supTopTit={'축제소개'}/>
-        <ul className="oneDep flex subDefaultContent">
-         <li className={oneDep === 1 ? "oneDeptit check" : "oneDeptit"} onClick={() => clickOne(1)}> <Link to='/Kinfo'>경복궁 별빛야행</Link></li>
-         <li className={oneDep === 2 ? "oneDeptit check" : "oneDeptit"} onClick={() => clickOne(2)}><Link to='/Cinfo'>창덕궁 달빛기행</Link></li>
-        </ul>
-        <ul className="twoDep flex">
-        <li className={twoDep === 1 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(1)}><Link to='/Kinfo'>축제정보</Link></li>
-        <li className={twoDep === 2 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(2)}><Link to='/Kprogram'>프로그램</Link></li>
-        <li className={twoDep === 3 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(3)}><Link to='/Kticket'>예매정보</Link></li>
-        </ul>
+        <InfoDep one={1}/>
+        <InfoDep2 one={1}/>
         <LineTit subtit={'축제정보'}/>
         <SubFesInfoK/>
         </>
@@ -40,29 +20,12 @@ function Kinfo(){
 }
 
 function Cinfo(){
-  const [oneDep, setOneDep] = useState(2);
 
-  const [twoDep, setTwoDep] = useState(1);
-    
-  function clickOne(index){
-      setOneDep(index);
-  }
-
-  function clickTwo(index){
-    setTwoDep(index);
-}
     return(
         <>
         <SupTop supTopImg={'subFesInfoImage'} supTopTit={'축제소개'}/>
-        <ul className="oneDep flex subDefaultContent">
-         <li className={oneDep === 1 ? "oneDeptit check" : "oneDeptit"} onClick={() => clickOne(1)}> <Link to='/Kinfo'>경복궁 별빛야행</Link></li>
-         <li className={oneDep === 2 ? "oneDeptit check" : "oneDeptit"} onClick={() => clickOne(2)}><Link to='/Cinfo'>창덕궁 달빛기행</Link></li>
-        </ul>
-        <ul className="twoDep flex">
-        <li className={twoDep === 1 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(1)}><Link to='/Cinfo'>축제정보</Link></li>
-        <li className={twoDep === 2 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(2)}><Link to='/Cprogram'>프로그램</Link></li>
-        <li className={twoDep === 3 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(3)}><Link to='/Cticket'>예매정보</Link></li>
-        </ul>
+        <InfoDep one={2}/>
+        <InfoDep3 one={1}/>
         <LineTit subtit={'축제정보'}/>
         <SubFesInfoC/>
         </>
@@ -70,29 +33,12 @@ function Cinfo(){
 }
 
 function Kpro(){
-  const [oneDep, setOneDep] = useState(1);
-
-  const [twoDep, setTwoDep] = useState(2);
-    
-  function clickOne(index){
-      setOneDep(index);
-  }
-
-  function clickTwo(index){
-    setTwoDep(index);
-}
+  
   return(
     <>
     <SupTop supTopImg={'subFesInfoImage'} supTopTit={'축제소개'}/>
-    <ul className="oneDep flex subDefaultContent">
-         <li className={oneDep === 1 ? "oneDeptit check" : "oneDeptit"} onClick={() => clickOne(1)}> <Link to='/Kinfo'>경복궁 별빛야행</Link></li>
-         <li className={oneDep === 2 ? "oneDeptit check" : "oneDeptit"} onClick={() => clickOne(2)}><Link to='/Cinfo'>창덕궁 달빛기행</Link></li>
-        </ul>
-        <ul className="twoDep flex">
-        <li className={twoDep === 1 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(1)}><Link to='/Kinfo'>축제정보</Link></li>
-        <li className={twoDep === 2 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(2)}><Link to='/Kprogram'>프로그램</Link></li>
-        <li className={twoDep === 3 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(3)}><Link to='/Kticket'>예매정보</Link></li>
-        </ul>
+    <InfoDep one={1}/>
+    <InfoDep2 one={2}/>
     <LineTit subtit={'프로그램'}/>
     <ProgramK/>
     </>
@@ -100,29 +46,12 @@ function Kpro(){
 }
 
 function Cpro(){
-  const [oneDep, setOneDep] = useState(2);
-
-  const [twoDep, setTwoDep] = useState(2);
-    
-  function clickOne(index){
-      setOneDep(index);
-  }
-
-  function clickTwo(index){
-    setTwoDep(index);
-}
+ 
   return(
     <>
     <SupTop supTopImg={'subFesInfoImage'} supTopTit={'축제소개'}/>
-    <ul className="oneDep flex subDefaultContent">
-         <li className={oneDep === 1 ? "oneDeptit check" : "oneDeptit"} onClick={() => clickOne(1)}> <Link to='/Kinfo'>경복궁 별빛야행</Link></li>
-         <li className={oneDep === 2 ? "oneDeptit check" : "oneDeptit"} onClick={() => clickOne(2)}><Link to='/Cinfo'>창덕궁 달빛기행</Link></li>
-        </ul>
-        <ul className="twoDep flex">
-        <li className={twoDep === 1 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(1)}><Link to='/Cinfo'>축제정보</Link></li>
-        <li className={twoDep === 2 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(2)}><Link to='/Cprogram'>프로그램</Link></li>
-        <li className={twoDep === 3 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(3)}><Link to='/Cticket'>예매정보</Link></li>
-        </ul>
+    <InfoDep one={2}/>
+    <InfoDep3 one={2}/>
     <LineTit subtit={'프로그램'}/>
     <ProgramC/>
     </>
@@ -130,29 +59,12 @@ function Cpro(){
 }
 
 function Kticket(){
-  const [oneDep, setOneDep] = useState(1);
-
-  const [twoDep, setTwoDep] = useState(3);
-    
-  function clickOne(index){
-      setOneDep(index);
-  }
-
-  function clickTwo(index){
-    setTwoDep(index);
-}
+ 
   return(
     <>
     <SupTop supTopImg={'subFesInfoImage'} supTopTit={'축제소개'}/>
-    <ul className="oneDep flex subDefaultContent">
-         <li className={oneDep === 1 ? "oneDeptit check" : "oneDeptit"} onClick={() => clickOne(1)}> <Link to='/Kinfo'>경복궁 별빛야행</Link></li>
-         <li className={oneDep === 2 ? "oneDeptit check" : "oneDeptit"} onClick={() => clickOne(2)}><Link to='/Cinfo'>창덕궁 달빛기행</Link></li>
-        </ul>
-        <ul className="twoDep flex">
-        <li className={twoDep === 1 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(1)}><Link to='/Kinfo'>축제정보</Link></li>
-        <li className={twoDep === 2 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(2)}><Link to='/Kprogram'>프로그램</Link></li>
-        <li className={twoDep === 3 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(3)}><Link to='/Kticket'>예매정보</Link></li>
-        </ul>
+    <InfoDep one={1}/>
+    <InfoDep2 one={3}/>
     <LineTit subtit={'예매 안내'}/>
     <TicketK/>
     </>
@@ -160,29 +72,12 @@ function Kticket(){
 }
 
 function Cticket(){
-  const [oneDep, setOneDep] = useState(2);
-
-  const [twoDep, setTwoDep] = useState(3);
-    
-  function clickOne(index){
-      setOneDep(index);
-  }
-
-  function clickTwo(index){
-    setTwoDep(index);
-}
+ 
   return(
     <>
     <SupTop supTopImg={'subFesInfoImage'} supTopTit={'축제소개'}/>
-    <ul className="oneDep flex subDefaultContent">
-         <li className={oneDep === 1 ? "oneDeptit check" : "oneDeptit"} onClick={() => clickOne(1)}> <Link to='/Kinfo'>경복궁 별빛야행</Link></li>
-         <li className={oneDep === 2 ? "oneDeptit check" : "oneDeptit"} onClick={() => clickOne(2)}><Link to='/Cinfo'>창덕궁 달빛기행</Link></li>
-        </ul>
-        <ul className="twoDep flex">
-        <li className={twoDep === 1 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(1)}><Link to='/Cinfo'>축제정보</Link></li>
-        <li className={twoDep === 2 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(2)}><Link to='/Cprogram'>프로그램</Link></li>
-        <li className={twoDep === 3 ? "twoDeptit tcheck" : "twoDeptit"} onClick={()=>clickTwo(3)}><Link to='/Cticket'>예매정보</Link></li>
-        </ul>
+    <InfoDep one={2}/>
+    <InfoDep3 one={3}/>
     <LineTit subtit={'예매 안내'}/>
     <TicketC/>
     </>
